FirebaseAuthentication is working but there is problems though. When I create an account, it creates an User in my real-time Database and the FirebaseAuth User aswell. On all the fragments everything is fine except that we will not use the feature to delete the tables, because it is now on firebase
When I try to login, it is successfull with FireBaseAuth, but I need the ID of the user from the realtime database so, I can't go further.
I made it work to see the tracking fragment but it is empty because of the problem I just explained in the last line.

My problem could be solved if I could take the real-time Database ID when I log with FirebaseAuth, but I have other things to do for school right now so I will not go further..